pyproject.toml
[project]
name = "gemini-ui-navigator-agent"
version = "1.0.0"
description = "Voice-driven web browser agent powered by Gemini Live API vision on Google Cloud"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.44.0",
    "google-cloud-aiplatform>=1.50.0",
    "playwright>=1.40.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
